Null-check for greeting members
For context: It would attempt to pull a config as soon as it joined the server if the server had random welcome messages, but that was before it had time to cache the server and generate a config for it to pull in the first place. Keep in mind there's no guarantee this race condition wouldn't still happen in the database, but I digress.