Potential fix, read desc
This hopefully fixes an issue where D#+ would throw ArgumentException, clamining the message does not contain any buttons, even though it does. I don't know *when* this became an issue, but I've already pointed it out to Cabbage, who's somewhat of the main maintainer of the lib for the time being. It's a regression in behavior, and I'll be testing this plugin for solutions.