[build] Overhaul to role menus, pt. 1
There's now the ability to edit the *options* of a role menu that's being configured.

This commit specifically improves some logic and allows for the re-use of CancellationTokens, which are set to expire ~vaguely~ when the initial interaction does, should someone wait that long. There's no logic to handle that, yet, and there may not be for a while, but it's better than nothing.

The next series of commits will aim to implement the ability to retroactively edit an existing role menu.