More help related fixes, fun.
There was an issue where tokens were in fact being properly consumed, but the formatter invoked the wrong method, which was not set up to handle single commands.

This has since been fixed, in regards to both issues.

Now, there is a check to see if the command being returned from `GetCommands` is in fact a child node (a command) and not a parent node (a group), and conditonially invokes the correct method.

Prior behavior: !help config aaa ➜ "Help for config edit"

New behavior: !help config aaa ➜ "Help for config edit aaa"