Add helper methods and responder for greeter
- Add CanAccessGreetingChannelAsync
This method determines whether or not a channel greeting is set, and if the member can access said channel.

- Add HasGreetingRoleAsync
Thie method determines whether or not a role-locked greeting is enabled, and whether a member has that role.

- Added a chcek to see if creating a new greeting fails, returning that error if it does.

- Added checks to ensure that greetings only evaluate if the member was queued for a particular reason (e.g. gained access to a channel) and a greeting exists to handle that.

- Added MemberGreetingResponder
This is a responder that responds to member adds, role updates, and channel updates.

If a member is added to a guild, it is immediately enqueued regardless, because the greeting service will return a non-successful result if greetings are not enabled in the guild.

If a member is updated, their roles are checked (first checking from cache), and if their roles are unchanged, the method returns.

if the member's roles are changed (or they were not previously in cache, perhaps offline), then HasGreetingRole is called, and they're optionally enqueued to the greeting service if they are.

In the event a channel is updated, pain.

This entirely relies on cache, which is generally fine, but may prove problematic in the future because Remora's caching scheme has some issues.

We attempt to get the channel from cache, and then we compare the overwrites from the current channel (in the context of the responder) with what was previously cached, and if there is one for a member, they're fetched from the API (which may fetch from cache; it's fine),

I've yet to validate any of this, but some external sources have verified the logic *appears* fine. Time to watch it blow up, I suppose.