Backend overhaul
There's *far* too much to document here, and it's all semi-related so I put it in one commit.

The gist of it is that the backend now uses `Snowflake`s instead of `ulong`s.

There's also some other consistency changes, and some documented things that were previously barren, as well as ensuring both tables and columns are set independently of their property's name