Add support sibling nodes
Adds support in the CommandHelpViewer and HelpFormatter to recognize when a group has a sibling node that shares the same name, effectivelly allowing said node to enable the facade that the group itself is being run. In other libraries this is usually refered to as a 'GroupCommand' or "Default group command".