🐛 FINALLY FIX CACHING ODDITIES
THIS HAS PLAGUED ME FOR MONTHS AND I FINALLY FIGURED IT OUT

SLIDING EXPIRATIONS SUCK. DO NOT USE THEM.
 Well, they suck when you expect absolute expirations to do anything.
I tried for months to fix this on various occasions, and I was ready to pull out Redis until I finally figured out that sliding expirations are the absolute devil :)

Absolute expiration means NOTHING if you're using a sliding expiration. Though it if we set the absolute expiration to null, we can get a cool effect of essentially dumping LRU objects every X hours, while extending the lifetime of objects we *do* need. This will hopefully fix a lot of issues, and will be a non-issue when this PR gets merged: https://github.com/Nihlus/Remora.Discord/pull/165

I can NOT express the reliquishment I felt when I finally narrowed down the actual cause of this issue. I've already brought it to the attention of the BCL team, as this behavior is somewhat unintuitive (sure, it makes sense if you actually access the object, but grrr)

A link to the issue is here: https://github.com/dotnet/runtime/issues/66445