🐛 Prefix regex should only match prefixes
Previously, it would match the prefix, but also any text trailing after it. This commit fixes this behavior by removing the greedy evaluation of any non-newline character, and also fixing a potential bug where a trailing space would cause the prefix to match, but not a command, due to how Remora's command parsing works by default.