[build core] 🐛 Assume the current year when parsing time
Fixes an issue where System.Text.Recognizers would return the previous year (if the time is technically in the future) when given a date without a year (e.g. April 13th). This would be the first result when parsing time, and cause an erroneous error stating that the reminder was being set in the past, when it very much wasn't.