🐛 Selecting roles properly updates the message now
Prior to, it would silently fail because Discord refsues to let you set 25 options as the maximium while keeping an optional minimum. This may be a bug in Remora? but I doubt it. Too tired too look into it at the moment, but if someone is so inclined to, then be my guest.