🐛 Allow users to select up to d options
This fixes a bug where selecting a role-menu would lock into only only allowing one role to be picked from that rolemenu after it updated.

Now, users can select up to d options, where d is the count of the options that already exist on the rolemenu.

This also adds a slight optimization of reusing the inital dropdown selected from `GetDropdownFromMessage`, and hopefully makes that code slightly more readable