🐛 Slash reminders are always private
Fixes a small bug where sending a slash command reminder in a server would consider it a public reminder because GuildID was set, and used as the discriminator.

This is remidied by checking if the message ID is null, which is indicative that the reminder was set from a slash command/modal, and thusly should always be considered private.